+++ src/game_api/level_api.cpp
@@ -1217,13 +1217,6 @@ void spawn_room_from_tile_codes(LevelGenData* level_gen_data, int room_idx_x, in
 using TestChance = bool(LevelGenData**, std::uint32_t chance_id);
 TestChance* g_test_chance{nullptr};
 
-struct SpawnInfo
-{
-    void* ptr0;
-    void* ptr1;
-    float x;
-    float y;
-};
 bool handle_chance(SpawnInfo* spawn_info)
 {
     auto level_gen_data = State::get().ptr()->level_gen->data;
@@ -1743,45 +1736,52 @@ uint16_t LevelGenData::get_pretend_room_template(std::uint16_t room_template)
     }
 }
 
+uint32_t ThemeInfo::get_aux_id()
+{
+    thread_local const LevelGenSystem* level_gen_system = State::get().ptr_local()->level_gen;
+    for (size_t i = 0; i < std::size(level_gen_system->themes); i++)
+    {
+        if (level_gen_system->themes[i] == this)
+        {
+            return uint32_t(i);
+        }
+    }
+
+    return uint32_t(~1);
+}
+
 void LevelGenSystem::init()
 {
     data->init();
+}
 
-    uint32_t real_id = 0;
-    for (ThemeInfo* theme : themes)
+void LevelGenSystem::populate_level_hook(ThemeInfo* self, uint64_t param_2, uint64_t param_3, uint64_t param_4, PopulateLevelFun* original)
+{
+    post_room_generation();
+
     {
-        theme->set_aux_id(++real_id);
-
-        using PopulateLevelFun = void(ThemeInfo * self, uint64_t param_2, uint64_t param_3, uint64_t param_4);
-        hook_vtable<PopulateLevelFun, 0xd>(
-            theme, [](ThemeInfo* self, uint64_t param_2, uint64_t param_3, uint64_t param_4, PopulateLevelFun* original)
-            {
-                post_room_generation();
-
-                {
-                    std::lock_guard lock{g_extra_spawn_logic_providers_lock};
-                    for (ExtraSpawnLogicProviderImpl& provider : g_extra_spawn_logic_providers)
-                    {
-                        provider.transient_num_remaining_spawns_frontlayer = provider.num_extra_spawns_frontlayer;
-                        provider.transient_num_remaining_spawns_backlayer = provider.num_extra_spawns_backlayer;
-                    }
-                }
-
-                original(self, param_2, param_3, param_4); });
-        using DoProceduralSpawnFun = void(ThemeInfo*, SpawnInfo*);
-        hook_vtable<DoProceduralSpawnFun, 0x33>(
-            theme, [](ThemeInfo* self, SpawnInfo* spawn_info, DoProceduralSpawnFun* original)
-            {
-                push_spawn_type_flags(SPAWN_TYPE_LEVEL_GEN_PROCEDURAL);
-                OnScopeExit pop{[]
-                                { pop_spawn_type_flags(SPAWN_TYPE_LEVEL_GEN_PROCEDURAL); }};
-
-                if (handle_chance(spawn_info))
-                {
-                    return;
-                }
-                original(self, spawn_info); });
+        std::lock_guard lock{g_extra_spawn_logic_providers_lock};
+        for (ExtraSpawnLogicProviderImpl& provider : g_extra_spawn_logic_providers)
+        {
+            provider.transient_num_remaining_spawns_frontlayer = provider.num_extra_spawns_frontlayer;
+            provider.transient_num_remaining_spawns_backlayer = provider.num_extra_spawns_backlayer;
+        }
     }
+
+    original(self, param_2, param_3, param_4);
+}
+void LevelGenSystem::do_procedural_spawn_hook(ThemeInfo* self, SpawnInfo* spawn_info, DoProceduralSpawnFun* original)
+{
+    push_spawn_type_flags(SPAWN_TYPE_LEVEL_GEN_PROCEDURAL);
+    OnScopeExit pop{[]
+                    { pop_spawn_type_flags(SPAWN_TYPE_LEVEL_GEN_PROCEDURAL); }};
+
+    if (handle_chance(spawn_info))
+    {
+        return;
+    }
+
+    original(self, spawn_info);
 }
 
 std::pair<int, int> LevelGenSystem::get_room_index(float x, float y)
