+++ src/injected/ui.cpp
@@ -55,6 +55,7 @@
 #include "decode_audio_file.hpp"
 
 #include "render_api.hpp"
+#include "script/lua_vm.hpp"
 #include "script/usertypes/vanilla_render_lua.hpp"
 
 #pragma warning(disable : 4366)
@@ -8008,6 +8009,10 @@ std::string make_save_path(std::string_view script_path, std::string_view script
 void init_ui()
 {
     g_SoundManager = std::make_unique<SoundManager>(&LoadAudioFile);
+
+    State::init();
+    get_lua_vm(g_SoundManager.get());
+    State::post_init();
 
     g_state = State::get().ptr_main();
     g_save = UI::savedata();
