+++ src/spel2_dll/spel2.h
@@ -58,6 +58,8 @@ class SpelunkyConsole;
 
 void Spelunky_SetDoHooks(bool do_hooks);
 void Spelunky_SetWriteLoadOptimization(bool write_load_opt);
+void Spelunky_InitState();
+void Spelunky_PostInitState();
 
 void Spelunky_RegisterApplicationVersion(const char* version);
 void Spelunky_InitMemoryDatabase();
